requirements.txt
psycopg2-binary==2.7.4
requests==2.22.0
redis==2.10.6
